api/requirements.txt
firebase-admin
google-cloud-firestore
Flask
tensorflow
Keras
h5py
opencv-python==3.3
requests
